+++ lib/statistrano.rb
@@ -8,6 +8,7 @@ require 'slugity/extend_string'
 require 'benchmark'
 
 # utility modules
+require 'statistrano/shell'
 require 'statistrano/git'
 require 'statistrano/log'
 require 'statistrano/ssh'
@@ -34,7 +35,7 @@ module Statistrano::DSL
 
     begin
       @deployment = ::Statistrano::Deployment.const_get(type.to_s.capitalize).new( name )
-    rescue NameError
+    rescue NameError => e
       ::Statistrano::LOG.error "The deployment type '#{type}' is not defined"
     end